Re-organize repo a bit & improve code clarity of React template
