+++ bin.js
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 
-require('./dist/lib/cli.js');
+require('./dist/cli/index.js');
