+++ cli/utils/merge-prompts.ts
