+++ scaffolds/hello-world-react/scaffold.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Template, Zombi } from 'zombi';
-import { createScaffold } from 'lib/utils/scaffold-helpers';
-import { NpmClientPrompt, PublicApiKeyPrompt } from 'lib/utils/common-prompts';
-import { mergePrompts } from 'lib/utils/merge-prompts';
+import { createScaffold } from 'cli/utils/scaffold-helpers';
+import { NpmClientPrompt, PublicApiKeyPrompt } from 'cli/utils/common-prompts';
+import { mergePrompts } from 'cli/utils/merge-prompts';
 
 type HelloWorldData = NpmClientPrompt.Data & PublicApiKeyPrompt.Data;
