+++ scaffolds/hello-world-react/template/src/components/Callback.js
@@ -1,6 +1,7 @@
 import React, { useEffect } from "react";
 import { useHistory } from "react-router";
-import { magic } from '../magic';
+import { magic } from "../magic";
+import Loading from './Loading';
 
 export default function Callback() {
   const history = useHistory();
@@ -12,10 +13,6 @@ export default function Callback() {
     });
   }, []);
 
-  return (
-    <div className="container">
-      <p>Loading...</p>
-    </div>
-  );
+  return <Loading />;
 }
