+++ scaffolds/hello-world-react/template/src/components/Loading.js
@@ -1,0 +1,10 @@
+import React from "react";
+
+export default function Loading() {
+  return (
+    <div className="container">
+      <p>Loading...</p>
+    </div>
+  );
+}
+
