+++ scaffolds/hello-world-react/template/src/index.html
@@ -2,7 +2,7 @@
 <html>
   <head>
     <meta charset="utf-8" />
-    <title>foobar</title>
+    <title>Magic Hello World 🌎<</title>
   </head>
 
   <body>
